Update license checker to handle Modifications copyright (#3)
Summary: We will have some code with a copyright on the modifications.
In those cases, we want to keep the original copyright authors but
include a line afterwards with our new modifications copyright.
If the license is already present in the file and it is Apache 2.0, then
we check the author of the copyright. If the author is the company, we
accept the license. Otherwise, we check the next line. If the next line
isn't the modifications copyright, we add it in.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Create files with different copyright configurations.

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 33b286d4c616839353de4ee7a5f4c1aa6d9a82a1