[lint] Speed up licenses checker by accepting list of files (#61)
Summary: `cli_lint_mode: file` is really really slow with mega-linter.
This changes the license checker to accept a list of files instead of a
single file. Takes the time to run the license checker from 7s to 0.1s.

Type of change: /kind infra.

Test Plan: Tested with `npx mega-linter-runner`.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: b2d2eb15c33cfc7a35584919a6795193196875b1