Add Terraform to create dev cluster (#47)
Summary: We are adopting a model where all devs share a single GKE
cluster. Devs are limited to modifying resources (pods, deployments,
etc) within their namespace only. This PR creates a dev cluster
(currently single-node, since we are not actively using it), and a
namespace per developer with the correct role + role-bindings. The list
of devs is currently hard-coded, but in the future we can add handling
to fetch the list of devs from Google Workspace.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Run `terraform apply`, connect to the cluster with `kubectl`,
then use `kubectl auth can-i` to test namespace permissions.

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
Signed-off-by: Zain Asgar <zasgar@gimletlabs.ai>
Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
Co-authored-by: Zain Asgar <zasgar@gimletlabs.ai>
Co-authored-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 8f5773abb18a0c98366d2fdc2919e5f3d090d692