[js_gazelle] Include @types/node by default (#83)
Summary: Adds a patch to js gazelle plugin, to support a gazelle
directive that specifies default types packages to include in all ts
rules. This is useful in particular for `@types/node` which is usually
included by default be `next` or other build tools, but needs to be
explicitly listed in the rules.

Type of change: /kind cleanup

Test Plan: Tested `bazel run //:gazelle` saw that the apollo client's
rule depends on `@types/node`.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: a8e4e520c5cae4029821a623c65361a178871be7