Add terraform to dev docker image (#137)
Summary: We want to start using terraform in one of the GH actions, to
check that the terraform has already been applied to the global state.
This adds tf to the docker image.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Run the docker image, check that terraform is there

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 230c446f47c63a703a7c179ff9d0a1124a8b634a