[build] Fix rules_python when run under root (#159)
Summary: rules_python, requires setting a flag on the toolchain to
enable running under root. This diff sets that flag so that CI can run
bazel build.

Type of change: /kind infra

Test Plan: Tested in a root docker container. Saw that it fails without
the change and works with it.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 2c6b6d3758042817793e4d6464d8db9c12256c4f