[go] Use pie by default on all go binaries (#218)
Summary: There's an issue with rules_go and sysroots where the stdlib is
compiled with absolute paths to the sysroot (rules_go claims that this
is necessary to avoid issues with go's sandbox while building the
stdlib). This means the stdlib .a archives have a cgo directive that
looks like `["cgo_ldflags",
"--sysroot=/path/to/stdlib/bazel/sandbox/rel/path/to/sysroot"]`. The
golang `link` tool reads the cgo directives from all of the archives its
linking, and parses them into flags for the external linker (i.e.
clang). It then passes these args to clang when determining if the
`-no-pie` flag is supported. So it ends up passing
`--sysroot=rel/path/to/sysroot` and
`--sysroot=/abs/path/to/sysroot/in/stdlib/sandbox` to clang, and clang
chooses the latter. Since the stdlib sandbox is no longer around, clang
can't find the sysroot and reports `-no-pie` as unsupported. If you run
in `--sandbox_debug` everything works just fine because the sandbox is
still around. To workaround this issue for now, this PR sets all
binaries to use PIE.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Tested running a go binary built with
`--config=x86_64_sysroot`.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 4d526efef4ea400b84d15c8f8f4b448e76e49ce4