[sysroots] Add feature flags to sysroots (#304)
Summary: We want to gate some sysroots features behind flags, allowing
us to avoid always downloading sysroots. This is the first step in
enabling that. This diff changes the sysroot creator to accept a
directory of features, where each feature is a directory of package
yamls for each variant of sysroot (runtime, build, or test). This diff
also moves the debug sysroot to be a "feature", so that it can also be
gated behind a flag in the future. The bazel changes to incorporate the
new feature flagged sysroots will come in a later diff, as well as
adding `cuda` as a new sysroot feature.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Ran `make build_all_sysroots` saw that it produced the debug
sysroot as a test variant as expected.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: d26a06d1f60e07e3a74e4a157365c5ede0328172