Fix UI build (#357)
Summary: This PR fixes the UI build problem causing errors like:
```
find: '.next/standalone/node_modules/.aspect_rules_js': No such file or directory
find: '.next/standalone/node_modules': No such file or directory
mv: cannot stat '.next/standalone/*': No such file or directory
mv: cannot stat '.next/standalone/.next': No such file or directory
mv: cannot move '.next/static' to 'out/app/.next/static': No such file or directory
```

There were two problems. First, when `next.js` is running in
`standalone` mode it traces all the files required by the server and
copies them over to the `standalone` directory. To do this it stores the
"tracing root", and calculates the traced files relative to this root.
By default this tracing root should be the directory next is run in.
However, next has an extra step where it looks for a lock file (any of
`pnpm-lock.yaml`, `yarn.lock`, etc) in the current directory and then up
through parent directories, if it finds one it sets the tracing root to
be the directory containing the lock file. When `next` was running
through bazel, the `pnpm-lock.yaml` was not included in the sandbox next
was running in. It then traversed up the directory tree looking for
other lock files, and since Z (maybe others as well) has a
`~/yarn.lock`, it decides the tracing root should be `~/` (bazel seems
to include `~/` as readonly in the sandbox not really sure why). So it
ends up putting all of the files in
`standalone/.cache/bazel/_...../src/ui/<files>`. The tar step then can't
find the files where it expects them and hence the above error.

The second issue is the lack of `set -e` in the tar step bash. Since the
above described errors didn't actually cause an exit code, the empty
generated tar was put into the cache. So for people without
`~/yarn.lock`, this bug appeared to be a cache issue because the cache
was polluted by people building with a lock file in their home dir.

Finally, we use an empty `pnpm-lock.yaml` instead of just including the
`pnpm-lock.yaml` directly because inside our bazel build `next.js`
decides that the `pnpm-lock.yaml` is a src module that it should try to
parse, and obviously fails. I haven't looked into this that much, I
suspect that there should be a way to include our actual
`pnpm-lock.yaml` but since next.js only seems to look for the presence
of a lock file, an empty file is fine for now.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Ran `touch ~/yarn.lock`. Then ran `bazel build
//src/ui:standalone` and saw the above described bug. Then applied this
diff and no longer saw the bug.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: aa202778d973dd87ef7cafe644ae429b9a4b9e8b