Add idempotency txn function (#366)
Summary: We would like to make sure that mutations (updates, creates)
are idempotent. For example, if the UI sends two of the same create
requests for some reason, we should not create two resources. Instead,
the UI should send an idempotency key. When running that query in our
DB, we should first make sure that the idempotency key does not already
exist. If it does, then we should reject the request. This PR adds a
helper function which users should use in any idempotent requests.

Relevant Issues: GML-113

Type of change: /kind feature

Test Plan: Unit tests

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 7bb45d92d0d55ab7980e3d9e5f7c0c68204737df