Add dev auth secrets to skaffold (#394)
Summary: We would like to deploy our secrets via skaffold. As stated in
#393, the Helm secrets functionality in Skaffold is currently broken. In
the meantime, we can use Kustomize to load in our SOPS'd secrets in
Skaffold. This PR adds the SOPS encrypted directory in the controlplane
manifests for the dev auth secrets. Alternatives would be to add the
kustomize manifests to the `infra/credentials` directory so that all of
our SOPS secrets are in one place.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: skaffold run -p dev

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 4aeee4ac4981dc1c2c39a726636b2da8aa8d4ad0