Add GetUserOrgs to directory service proto (#396)
Summary: Once a user signs up, we will need to check whether they are in
any orgs, in order to prompt them to create a new org. This adds the
"GetUserOrgs" call to the directory service where we can return a list
of orgs that a user is in, given their ID.

Relevant Issues: GML-73

Type of change: /kind feature

Test Plan: None, this just adds to the proto but doesnt use it anywhere
yet.

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 49f2b4655abf9c5a83e33fc1c0c1eabdb5d6e014