Implement CreateUser in directory service (#390)
Summary: This PR implements CreateUser in the directory service. A few
things to note: the service expects the userToken to be passed in via
the context. If the email of the user token does not match that of the
user being created, it will complain. The user server also expects the
idempotency key to be specified as a header on the request, and will
reject the request if no key is specified.

Relevant Issues: GML-73

Type of change: /kind feature

Test Plan: Unit test

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: e03447cc19afef4b612e506c34d1a8695486d2d6