Fix server logging framework (#404)
Summary: The old approach was causing GRPC errors due to an issue in
mutil. This wraps http
with the writer interface which also removes a dependency.

Still going to improve the logging, especially for development.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Tested on skaffold deploy

Signed-off-by: Zain Asgar <zasgar@gimletlabs.ai>
GitOrigin-RevId: 06f0759c6e2bfecef42d15d3c34fbc2984f84ec4