Implement CreateOrg in Directory service (#427)
Summary: This implements CreateOrg in the directory service. This will
create the org, then add the user who created the org as the admin.
Since we are no longer passing the usertoken in the auth, we needed to
add the ID of the creator user in the proto.

Relevant Issues: GML-74

Type of change: /kind feature

Test Plan: Unit test

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 7b28556b0a826f4bddbaf7ac88f852511446daa9