[tensorflow fixes] Patch cpuinfo to support platforms (#464)
Summary: tensorflow depends on `pytorch/cpuinfo`, but the latter still
uses the old way of doing cross compiling in bazel (`--cpu` instead of
`--platforms`). This adds a patch to make it use the new way for the
platforms we care about.

Relevant Issues: GML-141

Type of change: /kind cleanup

Test Plan: Tested as part of tensorflow build with future PR.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 8dbe35754055b34733e02c8c4b84dc590d150061