Remove domainName from directory service (#526)
Summary: We aren't using domain name anywhere yet, and we had a unique
constraint on domainName which was breaking org creation. We should just
remove it and add it back later in the future if we ever decide to use
it.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: Skaffold run, unit tests

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 39ac6692c86714628a245f384be51b9877311f02