[mediapipe] Use our glog in mediapipe (#552)
Summary: Mediapipe was using a different version of glog from the
version we use in our code. This was causing duplicate symbol issues
when building binaries that use glog and mediapipe.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Tested with other changes that I can now build a binary that
uses both glog and mediapipe.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 05b0b078240de0ca19d2fd0a6c7bccd09f521a84