[gem] Add core::ExecutionContext (#554)
Summary: Currently, I've structured GEM as a series of plugins. Each
plugin can have custom mediapipe calculators. To support plugins that
require state, this PR adds `core::ExecutionContext` which is the base
class for any plugin `ExecutionContext`. `core` will call out to a
plugin to build its `ExecutionContext`, and then handle passing the
context pointer to the plugins calculators.

An example usage of an `ExecutionContext` is the `tensorrt` plugin (not
yet committed), which uses the `ExecutionContext` to store a `TensorRT`
engine.

Relevant Issues: GML-142.

Type of change: /kind feature

Test Plan: This PR is just a virtual base class, so no tests yet.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 05347554572abd515f1a1923ad73b09b28f7a3d2