[ffmpeg] Fix shared library names (#570)
Summary: The ffmpeg shared libraries need to be versioned. Otherwise,
binaries are linked such that they look for the versioned shared library
but bazel only provides the unversioned one.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Tested with a new test to be added in another PR. The test
failed before because it couldn't find libavcodec.so.58, but now it
works.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: d517a513e040d7abcffa486df1738776af9689bd