[gem] Argus cam device (#600)
Summary: This adds the Argus camera access functionality into the GEM.
It is currently stand-alone.

Relevant Issues: Towards GML-165

Type of change: /kind feature

Test Plan: Tested manually. This is difficult to test in our build
system, because it is specific to the Jetson device. In a follow-up, I
plan on including an example binary that can be deployed onto the device
as part of a manual test.

Changelog Message:

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: 13b0c1b70488c2cb250dc98864f15b0d39fc4d31