[gem] Add TensorRT ExecutionContextBuilder to build TensorRT engines. (#595)
Summary: Adds an `ExecutionContextBuilder` for the `TensorRT` plugin.
The builder currently takes a path to an onnx model file and builds a
`TensorRT` engine from it. It then wraps the engine in a
`tensorrt::ExecutionContext` for use in mediapipe calculators.

Relevant Issues: Fixes GML-146

Type of change: /kind feature

Test Plan: Still need to figure out a CI test plan for targets that
require tensorrt. For now, I tested it on a Jetson.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 4dc2c5a16cc61b7a3bd5521dfa313452686e29f5