Add and implement GetFleet in fleetmgr service (#611)
Summary: We will be using GetFleet to ensure that a user has permissions
to create a deploy key for a fleet. This adds the proto and implements
it in the fleetmgr service. We need to handle both if the request is
made by a service or user token.

Relevant Issues: Towards GML-177

Type of change: /kind feature

Test Plan: unit test

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 86184e194e336c50abe6b3c20e0876eec344f961