[gem] Add cpu_tensor plugin (#606)
Summary: Adds gem plugin that adds a CPU resident tensor. Currently, the
CPU tensor is just a memory buffer without any shape or data type
information. In the future, we probably want to use a library like
`Eigen` to represent CPU tensors, but this is a simple path forward for
now.

Relevant Issues: Towards GML-174.

Type of change: /kind feature

Test Plan: Beyond the CPUTensorPool which is tested as part of
core::TensorPool, there isn't really anything else to test in this diff.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 882f7939c723402211158764eb335b68cb5f9417