[gem] Argus cam test binary (#617)
Summary: This is an example binary that can be used for testing the
ArgusCam code on the Jetson device. Unfortunately, it cannot be built on
our existing build system, because of an incompatibility between our
arm64 sysroot and the Jetson device's sysroot. Nevertheless, checking it
in so we have it for future use.

Relevant Issues: Towards GML-165

Type of change: /kind test

Test Plan: This is a pseudo-test. It is a binary that can be run
manually on the Jetson device.

---------

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: d30e05c692a0a44a1eeb8999b09295b65182be09