Add serviceID check to ValidClaims() (#663)
Summary: I was repeating this check across all of the services multiple
times, and it seemed like something that should go into the
`ValidClaims` check instead, which already does half of the check. In
order to get the serviceID, I stick it into the authcontext in the
middleware.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: unit tests

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 70afd05115d591246b42c1015eee8b02564f27ae