ArgusCam: Expand NvBufSurfaceWrapper (#686)
Summary: This expands the functionality of the NvBufSurfaceWrapper with
some debug information and the ability to map the buffer for CPU access.
These have been useful in manual testing purposes.

Relevant Issues: Towards GML-165

Type of change: /kind feature

Test Plan: Manual testing on the Jetson device, and then adapting for
our repo.

GitOrigin-RevId: b1e543e99c1669b474093cd64340ac533a48d881