[gem] Add calculator to get the shape of an image as a tensor (#629)
Summary: Some models such as YOLOv3 require the original image shape as
an input. This PR adds a calculator that takes a mediapipe::ImageFrame
and outputs its shape as a CPUTensor.

Relevant Issues: Towards GML-174.

Type of change: /kind feature.

Test Plan: Added a test for the calculator using the framework from
#628.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: f80d838e73a524e582b4548d3ddc061beaa4b2bc