Bazel repos for ArgusCam: Pull in libnvsurface.so (#689)
Summary: This sets up our repo to pull libnvsurface.so, which is
required for building the ArgusCam on the Jetson device.

Relevant Issues: GML-165

Type of change: /kind feature

Test Plan: Manual testing that the build still works, and that fewer
symbols are missing when manually building the argus_cam_binary.

GitOrigin-RevId: e6b3fcbcb4371f094aea2d86b2acf85add348f8d