NvBufSurfaceWrapper: Remove dependency (#692)
Summary: A small simplification that removes a header dependency. We
might be able to get rid of the jetson_multimedia_api completely after
this, but leaving that stuff in just in case for now.

Relevant Issues: GML-165

Type of change: /kind feature

Test Plan: Jetson-specific device code, so manual testing for now.

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: cb69baf535b28f721b5810789887321a53e0ff10