[common] Add RAII wrapper around a linux file (#722)
Summary: Add a RAII wrapper around a linux file, so that the file
descriptor is automatically closed when the file goes out of scope.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Added a test for the wrapper that makes sure the file
descriptor is no longer open after the wrapper goes out of scope.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: e3098c99fcdec2a1d33a7451f9bce7f9059a1684