Add patches to allow paths with parens in our TS project (#732)
Summary: Nextjs uses folders with parens to do route grouping without
affecting the route. We'd like to use the same and so these patches
help support parens in other tooling we use.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Tried building and testing with these patches and a path with
parens.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: ac6848259b75c2dae90ea98f2ff61f5e8bac8566