[gem] Add BlobStore for storing binary data (#734)
Summary: Adds the abstract base class for BlobStores in GEM. BlobStores
store arbitrary binary blobs of data. The storage of the data and
mapping it into memory are left up to the implementations.

Relevant Issues: Towards GML-207.

Type of change: /kind feature

Test Plan: This PR just adds a base class, the next PR adds an
implementation and a corresponding test.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 8795674e924906ba5c680d65557fbbc7129ed659