[gem] Use cached TensorRT engine from BlobStore if available (#738)
Summary: Allow the `ModelSpec` to pass a blob key for a cached
`TensorRT` engine. If the key is passed in the spec, the model builder
will attempt to use the cached engine, and if it doesn't exist it will
build it and then write it back to the `BlobStore` for future use.

Relevant Issues: GML-207

Type of change: /kind feature

Test Plan: Tested on a jetson by running once to build the engine, saw
it was saved to a path inside my `FilesystemBlobStore`s directory. Then
ran it again and saw it loaded the cached engine without rebuilding it
in TensorRT.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 3e71856bde8cc17e6313c11148b884b66a314a0c