[gem] Implement PlanarImage for mediapipe's YUVImage (#775)
Summary: Adds a template `PlanarImageFor` that allows implementing the
`PlanarImage` interface as long as the `PlanesFromImage` funciton is
specialized for the given image type. This template is then used to
implement the interface for mediapipe's YUVImage.

Relevant Issues: Towards GML-206.

Type of change: /kind feature.

Test Plan: Added a test for the conversion from `mediapipe::YUVImage` to
`PlanarImage`.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: cb707f00d24eadb6ef2b4de51d982e8f04af7347