Mediapipe calculator for converting NvBufSurface to mediapipe::ImageFrame (#761)
Summary: This calculator uses libyuv to convert from the nvbufsurface to
a generic mediapipe::ImageFrame, which will make it usable by subsequent
model stages.

Relevant Issues: Resolves GML-202

Type of change: /kind feature

Test Plan: A test is included, but needs to be run manually on the
Jetson device. A container image is provided for that purpose. The test
output is not yet validated either, so there will likely be some
follow-on work.

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: dd8332d5f69e7ffa225becc77b70476631c72301