[gem] PlanarImage: Interface tweaks (#816)
Summary: There are a few of changes:
1) Part of the interface was specific to YUVImage, and this is now
decoupled.
2) Collapsed one layer of indirection between PlanesFromImage and
BuildPlanes. I'm okay keeping the extra layer too, but do want it to be
consistent with how the Height() and Width() are implemented (i.e. they
should go through a layer of indirection too if we want to go that
route).
3) Converted struct Plane::data to be a const pointer.

Relevant Issues: Towards GML-222

Type of change: /kind cleanup

Test Plan: Existing tests

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: 1b8c23360dee23361bb8bba6c0de312041786b57