Script to help run tests on Jetson device (#769)
Summary: This shell script is for convenience of deploying a test to run
on the Jetson device. Eventually this should be superceded by Bazel
executors that can deploy to a fleet of Jetsons. But for now this offers
some convenience.

Relevant Issues: Towards GML-205

Type of change: /kind testing

Test Plan: This is test infrastructure. Manually tested.

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: b15680d277e4663a6503a22d32c5c0ce04e04624