[gem] Implementation of PlanarImage for NvBufSurfaceWrapper (#828)
Summary: This implementation of PlanarImage for NvBufSurfaceWrapper will
enable us to hook up the ArgusCamCalucaltor to other calculators.

Relevant Issues: Towards GML-222

Type of change: /kind feature

Test Plan: Test is not functional because it requires Argus (which
doesn't work on x86), and the device specific test is missing a shared
library to become functional. The test will be submitted as follow-up
work.

Changelog Message:

GitOrigin-RevId: 36fb7f4ce097d2ec333704187d42e7de82fd74dd