Implementation of a video player that can function with h.264 NAL units (#885)
Summary: This implements the video player based on the broadway-player
library. Broadway itself does
not follow canonical React rules (since it's not based on react). This
reimplemts part of the player to make
it more react friendly by ensuring that new DOM elements are not created
as part of class and react is fully aware of all the DOM elements.

It also currently only supports WebGL. We might want to fix this if we
run into issues.

Relevant Issues: closes GML-93

Type of change: /kind feature

Test Plan: It's hard to implement tests for WebGL code. There is a new
story added that fetches a static video file based on NAL units.
Verified that the video renders properly.

GitOrigin-RevId: 5842ecdcd8b5a6323cad25007c82718cc7094ce7