[gql] Add DeployModelToFleet mutation (#921)
Summary: Sets up the DeployModelToFleet mutation in the api service. The
actual model deployment implentation is mocked out, but mostly matches
what we expect to happen in the future. This does not change any state,
that will come in a later PR.

I followed the conventions that we set with most of the other mutations
- use a single argument 'input'. This is recommended by apollo's
documentation.

Relevant Issues: Towards GML-189

Type of change: /kind feature

Test Plan: Added tests and also verified this works on the UI.

---------

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: c34171845b9eb6922a27c954297a7d2df441db8e