Add a WebSocket server util that serves the static bbox_video for testing (#945)
Summary: Refactors the testutils for bbox_video to share code and
adds a websocket server to serve the video and bbox data over a
socket.
Also adds a "test" and "update" rule to make sure that the json.gz
video_bbox file for storybook is up to date and easy to update.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Ran the websocket server manually.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: dd2c44aa5b3fa7cf052eca1ac7d3fd9b5f6df4c7