[common] A utility to extract MAC address from host (#894)
Summary: A utility to extract the MAC address from /sys/class/net. It
will ignore any virtual devices (like docker and br devices), but
otherwise use the lexicographically first device on the host.

Relevant Issues: GML-231

Type of change: /kind feature

Test Plan: A basic test runs the function. The test checks for physical
devices, so there is some risk that the test will behave differently on
different hosts. For example, if a host has no physical devices, this
test might fail. We can deal with that scenario if it ever comes to
light.

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: 9248859259f48d9ce39566d37e34fe204ef0168b