[controlplane/modelserver] Add stub modelserver. (#946)
Summary: Adds a stub for the modelserver. The main RPC for now is
`RunModel`. In the future, we expect to split the RPCs into something
like `SetupModel` and `Run` instead of a single call.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: This is just a stub, but it's Tested as part of the
implementation PR.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 2884f28fc57b5031de022cab88daea438ca8eb98