fleetmgr: Add tags to devices and API to update device tags (#959)
Summary: We want to be able to associate tags with devices, just like we
do for fleets. This PR adds tags to the devices table, and it also adds
an API `UpdateDevice` that allows the user to update the tags for a
given device. This is outside of the RegisterDevice flow because that is
a backend API that will not have access to device tags. Additionally,
the user may want to update the device tags once a device has been
registered.

Relevant Issues: Towards GML-223.

Type of change: /kind feature

Test Plan: added unit test, extended existing unit tests

GitOrigin-RevId: fab07c776fbffcf9e87ced147e073bee07500538