[common] MacAddress: Parameterize to accept the /sys/class/net location (#1028)
Summary: Parameterize the Mac Address reader code so that we can provide
an alternate path. This is important for when the code is running inside
a container. Assuming the host /sys/class/net is volume-mounted into the
container, then it becomes possible to still get a unique device ID by
reading the host MAC address.

Relevant Issues: Closes GML-244

Type of change: /kind feature

Test Plan: Tested to make sure it still works in the non-containerized
way. Hard to create automated test for the containerized environment, so
that is left out of scope.

GitOrigin-RevId: cf219e25853676e3671af906e9970ec12fabcf44