Create a bazel config for the Jetson (#1088)
Summary: With the merge of #1064, we have an additional flag we need to
set to build the Argus code, so create a Jetson config for convenience.
In the future, we should plan on having a Jetson sysroot, and actually
enabling this config in our CI/CD ppipeline.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: Manually tested.

GitOrigin-RevId: 0e3e352b7dfae049c369627e4803289bf35f9b46