[sysroots] Support multiple package database index files. (#1129)
Summary: Adds support for specifying multiple package indices to use for
any given sysroot. At this point, the package_satisfier has grown a bit
beyond what it was intended for, and has become a bit of a mess. I hope
to clean it up in the future but for now I will stick with the current
structure.

Relevant Issues: Towards GML-247

Type of change: /kind cleanup

Test Plan: Tested as part of making a jetson sysroot. Saw that it can
use both the ubuntu package index and the nvidia package index at the
same time.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: f1c295bd23ff6696ad0802216b113335c8f4bfac