[sysroots] Support version constraints in package_satisfier. (#1131)
Summary: Adds support for versioned dependencies in the package
satisfier. This is necessary for the jetson sysroot because NVIDIA deb
repos include multiple versions of relevant packages and we want to
match the version in Jetpack 5.1.2 exactly.

Relevant Issues: Towards GML-247.

Type of change: /kind clenaup

Test Plan: Tested when building the jetson sysroot. Saw that it can
select version 8.5.2 of tensorrt and install the correct corresponding
dependencies. Also checked that the current sysroots remain unchanged
with these changes.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: e084f4ebb90349ea5eabfe1afcdb7bb041d22cc7