[sysroots] Add Jetson sysroot (#1136)
Summary: Adds a sysroot targetting the Jetson for Jetpack 5.1.2.

Relevant Issues: Towards GML-247.

Type of change: /kind feature.

Test Plan: Built the sysroot. Along with the next PR that adds a clang
config for this sysroot, tested that I can build with this sysroot and
run the resulting binary on the jetson.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 98d6fd365a1ba71a8d39f48cf7c4fc3acc55ced7