[build] Fix clang tidy (#1141)
Summary: Adds a number of fixes to get clang-tidy working with our
sysroot build.
- Patches `clang-tidy-diff.py` with a backport from llvm main that
causes it to error if there are errors in clang-tidy. (This was hiding
the fact that clang-tidy was broken)
- Updates the minimal clang build to include the `clang-tidy` binary.
This is necessary because the `clang-tidy` binary uses its location to
find the correct clang headers to include.
- Fixes `ci/run_clang_tidy.sh` to only run `clang-tidy` on files that
are in the generated `compile_commands.json`. This prevents `clang-tidy`
from running on files that are incompatible with the x86_64 sysroot. In
the future, we can add support for other sysroots to
`gen_compilation_database` and then `clang-tidy` should run on the files
added by the new sysroot support.
- Add a patch to `bazel-compilation-database` that supports quotes in
defines. For example `GLOG_EXPORT` has `...("default")` in it. Before
the patch this was being converted to a compile command that would have
`GLOG_EXPORT=...(default)`. With the patch the quotes stay around so
that the compile command actually works.
- Change the `bazel` wrapper script to add the `x86_64_sysroot` config
by default for `cquery` and `aquery` where previously it only added it
for `build`, `test`, and `run`.
There are a bunch of clang tidy issues in our codebase, this PR doesn't
try to solve any of them. The next PR puts out a bunch of fixes for most
of the issues.

Relevant Issues: Fixes GML-250.

Type of change: /kind infra

Test Plan: Tested that `clang-tidy` runs without errors now.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 88bf7f16917a75983165de6b2656ede95a09033f