[build] Vendor mediapipe protos (#1158)
Summary: We want to be able to create GEM execution specs in the
controlplane. However, that requires that the protos work with golang.
In order to build go libs for mediapipe protos we need to vendor the
mediapipe protos. This PR vendors mediapipe protos into `third_party`.
The mediapipe protos, by default, are incompatible with golang, so a lot
of changes went into our fork of mediapipe to make the protos compatible
with golang. Most of the changes were prefixing enum values so that they
don't conflict with other enum values in the same go package. Because of
the effort required to get the protos working, I didn't vendor every
single proto from mediapipe. This means some parts of mediapipe will now
fail to build, and if we need to use those parts, we'll have to vendor
more of the protos. I believe the protos I've vendored here should be
sufficient for our purposes but I wouldn't be surprised if we need more
in the future.

Relevant Issues: Towards GML-252

Type of change: /kind cleanup

Test Plan: Should not be any functional changes, relying on existing
tests to verify that.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 7e00990172671ef365d24d654091c63c0c42c92a