[signal] Remove deadlock in fatal signal handler (#1209)
Summary: There's a possible deadlock in our stacktrace handler. When a
`CHECK` failure happens `glog` dumps the stacktrace and then aborts
(while holding the logging lock), then our signal handler triggers and
attempts to `LOG(ERROR)` causing a deadlock. I'm not really sure why
we're only seeing this deadlock with the jetson sysroot, but for now
this is an easy fix.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Tested with the `jetson_sysroot` with and without this
change. Saw that it deadlocks without it, and doesn't with this change.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: e2677f2cbd81cad9cc88e0a8944e4baa55054b77