[GEM-Controller] Add GRPC bridge controller and message handler interface (#1211)
Summary: This adds the GRPC bridge that handles connection to the
control plane. It also adds support for a message handler that can be
used to handle
message from the controlplane.

Relevant Issues: towards GML-229

Type of change: /kind feature

Test Plan: We need to figure out how to properly test bidi streams. We
can add tests after that.

Signed-off-by: Zain Asgar <zasgar@gimletlabs.ai>
GitOrigin-RevId: 6e56ba41bc86e37b4e0699a33728eddeb88906bc