Add a basic GEM install/setup script (#1208)
Summary: This adds a basic shell script to download and run the GEM.
This is just the shell, we need to add the image fetch and run
commands. The script also needs to be included in our controlplan as
a static resource that can be downloaded.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: Ran the script and checked the uname error path and the
deploy key prompt.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: d5dbdebacd1d78e8d7939188f32f0fa5b9ace63d