Add durable partition handler (#1219)
Summary: Now that our jetstream utils are in, this PR adds a
DurablePartitionHandler for handling partition messages from jetstream.
I opted for creating a completely separate partition handler rather than
combining it with the existing partition handler, since the subscription
models are slightly different (we don't need to throw the jetstream msg
on a channel + it needs to be acked).

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: unit test

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 9393cd2620bc3271a683ae6f34f116034b80fa5d