[jetson] Add gml_jetson_image to create oci image for jetson (#1199)
Summary: Adds `gml_jetson_image` to replace
`gml_host_cuda_binary_image`. It uses the runtime `jetson_sysroot` as
the base of the image, with the expectation that cuda/nvidia libs will
be mounted from the host at runtime. This PR also removes the hack we
had in place for including ffmpeg shared libraries in
`gml_host_cuda_binary_image`s, replacing it with a system that gathers
the runfiles from a binary and puts them in the image in the directory
structure bazel expects. As part of this, this PR cleans up some of the
tests that used `gml_host_cuda_binary_image` with hard-coded testdata
paths.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Tested that the `gml_host_cuda_binary_image` jetson tests now
work with `gml_jetson_image`.

GitOrigin-RevId: cb2e01e154d42fcf69f561a36d86f119f3b76b87