[gem] xvfb wrapper script to start fake x11 server (#1252)
Summary: Adds `Xvfb` to sysroot for running a virtual x11 server.
libargus requires EGL to operate, so we need an X11 server to talk to.
This adds an option to `gml_jetson_image` to enable starting a virtual X
server in the container before running the main binary.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Tested that the X11 server starts properly. Also tested that
libargus works with the virtual server.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: fe42bc68ed14a7b895a5f821316871e51f769ce9